fix typo, add date demo
